+++ pvl/encoder.py
@@ -16,6 +16,7 @@ import re
 import textwrap
 
 from collections import abc
+from warnings import warn
 
 from ._collections import PVLObject, PVLGroup, Units
 from .grammar import PVLGrammar, ODLGrammar, ISISGrammar
@@ -70,6 +71,52 @@ class PVLEncoder(object):
         self.end_delimiter = end_delimiter
         self.aggregation_end = aggregation_end
         self.newline = newline
+
+        # This list of 3-tuples *always* has our own pvl quantity object,
+        # and should *only* be added to with self.add_quantity_cls().
+        self.quantities = [(Units, 'value', 'units')]
+
+        try:
+            from astropy import units as u
+            self.add_quantity_cls(u.Quantity, 'value', 'unit')
+        except ImportError:
+            warn("The astropy library is not present, so "
+                 "astropy.units.Quantity objects will not "
+                 "be properly encoded.", ImportWarning)
+
+        try:
+            from pint import Quantity as q
+            self.add_quantity_cls(q, 'magnitude', 'units')
+        except ImportError:
+            warn("The pint library is not present, so "
+                 "pint.Quantity objects will not "
+                 "be properly encoded.", ImportWarning)
+
+    def add_quantity_cls(self, cls, value_prop: str, units_prop: str):
+        """Adds a quantity class to the list of possible
+        quantities that this encoder can handle.
+
+        :param cls: The name of a quantity class that can be tested
+            with ``isinstance()``.
+        :param value_prop: A string that is the property name of
+            *cls* that contains the value or magnitude of the quantity
+            object.
+        :param units_prop: A string that is the property name of
+            *cls* that contains the units element of the quantity
+            object.
+        """
+        if not isinstance(cls, type):
+            raise TypeError(f"The cls given ({cls}) is not a Python class.")
+
+        # If a quantity object can't encode "one meter" its probably not
+        # going to work for us.
+        test_cls = cls(1, 'm')
+        for prop in (value_prop, units_prop):
+            if not hasattr(test_cls, prop):
+                raise AttributeError(f"The class ({cls}) does not have an "
+                                     f" attribute named {prop}.")
+
+        self.quantities.append((cls, value_prop, units_prop))
 
     def format(self, s: str, level: int = 0) -> str:
         """Returns a string derived from *s*, which
@@ -224,12 +271,30 @@ class PVLEncoder(object):
         """Returns a ``str`` formatted as a PVL Value based
         on the *value* object according to the rules of this encoder.
         """
-        if isinstance(value, Units):
-            val = self.encode_simple_value(value.value)
-            units = self.encode_units(value.units)
-            return f'{val} {units}'
-        else:
+        try:
+            return self.encode_quantity(value)
+        except ValueError:
             return self.encode_simple_value(value)
+
+    def encode_quantity(self, value) -> str:
+        """Returns a ``str`` formatted as a PVL Value followed by
+        a PVL Units Expression if the *value* object can be
+        encoded this way, otherwise raise ValueError."""
+        for (cls, v_prop, u_prop) in self.quantities:
+            if isinstance(value, cls):
+                return self.encode_value_units(getattr(value, v_prop),
+                                               getattr(value, u_prop))
+
+        raise ValueError(f"The value object {value} could not be "
+                         "encoded as a PVL Value followed by a PVL "
+                         f"Units Expression, it is of type {type(value)}")
+
+    def encode_value_units(self, value, units) -> str:
+        """Returns a ``str`` formatted as a PVL Value from *value*
+        followed by a PVL Units Expressions from *units*."""
+        value_str = self.encode_simple_value(value)
+        units_str = self.encode_units(str(units))
+        return f'{value_str} {units_str}'
 
     def encode_simple_value(self, value) -> str:
         """Returns a ``str`` formatted as a PVL Simple Value based
@@ -618,7 +683,7 @@ class ODLEncoder(PVLEncoder):
         """
 
         # if self.is_identifier(value.strip('*/()-')):
-        if self.is_identifier(re.sub(r'[\*/\(\)-]', '', value)):
+        if self.is_identifier(re.sub(r'[\s\*/\(\)-]', '', value)):
 
             if '**' in value:
                 exponents = re.findall(r'\*\*.+?', value)